Move the registration of the node from calicoctl into the container.
Also
- Clean up passing in felix env vars
- Print out the docker command that is used to run the container
