+++ calico_node/filesystem/etc/rc.local
@@ -1,3 +1,5 @@
+/startup || exit 1
+
 if [ "$CALICO_DISABLE_FILE_LOGGING" == "true" ]; then
 	rm -r /etc/service/bird/log
 	rm -r /etc/service/bird6/log
@@ -7,7 +9,7 @@ fi
 
 if [ "$CALICO_NETWORKING" == "false" ]; then
 	# If running in policy only mode, we don't need to run BIRD / Confd.
-	echo "Running in policy-only mode - disable BIRD and Confd"
+	echo "CALICO_NETWORKING is false - disable BIRD and Confd"
 	rm -r /etc/service/bird
 	rm -r /etc/service/bird6
 	rm -r /etc/service/confd
@@ -24,8 +26,10 @@ else
     # running twice ensures our starting configuration is correct.
     confd -confdir=/etc/calico/confd -onetime -node=${ETCD_SCHEME:=http}://${ETCD_AUTHORITY} \
           -client-key=${ETCD_KEY_FILE} -client-cert=${ETCD_CERT_FILE} \
-          -client-ca-keys=${ETCD_CA_CERT_FILE} -scheme=${ETCD_SCHEME:=http} || true
+          -client-ca-keys=${ETCD_CA_CERT_FILE} -scheme=${ETCD_SCHEME:=http} >/felix-startup-1.log 2>&1 || true
     confd -confdir=/etc/calico/confd -onetime -node=${ETCD_SCHEME:=http}://${ETCD_AUTHORITY} \
           -client-key=${ETCD_KEY_FILE} -client-cert=${ETCD_CERT_FILE} \
-          -client-ca-keys=${ETCD_CA_CERT_FILE} -scheme=${ETCD_SCHEME:=http} || true
+          -client-ca-keys=${ETCD_CA_CERT_FILE} -scheme=${ETCD_SCHEME:=http} >/felix-startup-2.log 2>&1 || true
 fi
+
+echo "Calico node started successfully"
