+++ calico_node/filesystem/etc/service/felix/run
@@ -1,3 +1,8 @@
 #!/bin/sh
 exec 2>&1
+export FELIX_ETCDADDR=$ETCD_AUTHORITY
+export FELIX_ETCDSCHEME=$ETCD_SCHEME
+export FELIX_ETCDCAFILE=$ETCD_CA_CERT_FILE
+export FELIX_ETCDKEYFILE=$ETCD_KEY_FILE
+export FELIX_ETCDCERTFILE=$ETCD_CERT_FILE
 exec calico-felix
