+++ calico_node/start_runit
@@ -2,12 +2,11 @@
 # From https://github.com/faisyl/alpine-runit
 env > /etc/envvars
 
-echo >&2 "*** Running: rc.local"
 /etc/rc.local
 retval=$?
 if [ $retval -ne 0 ];
 then
-    echo >&2 "*** Failed with return value: $?"
+    echo >&2 "Calico node failed to start"
     exit $retval
 fi
