+++ calicoctl/tests/unit/node_test.py
@@ -57,7 +57,7 @@ class TestAttachAndStream(unittest.TestCase):
         m_stdout = Mock(spec=sys.stdout)
         m_sys.stdout = m_stdout
         m_container = Mock()
-        node._attach_and_stream(m_container)
+        node._attach_and_stream(m_container, False)
 
         m_docker_client.attach.assert_called_once_with(m_container,
                                                        stream=True)
@@ -85,7 +85,7 @@ class TestAttachAndStream(unittest.TestCase):
         m_stdout = Mock(spec=sys.stdout)
         m_sys.stdout = m_stdout
         m_container = Mock()
-        node._attach_and_stream(m_container)
+        node._attach_and_stream(m_container, False)
 
         m_docker_client.attach.assert_called_once_with(m_container,
                                                        stream=True)
@@ -116,7 +116,7 @@ class TestAttachAndStream(unittest.TestCase):
         m_stdout = Mock(spec=sys.stdout)
         m_sys.stdout = m_stdout
         m_container = Mock()
-        node._attach_and_stream(m_container)
+        node._attach_and_stream(m_container, False)
 
         m_docker_client.attach.assert_called_once_with(m_container,
                                                        stream=True)
@@ -177,10 +177,6 @@ class TestNode(unittest.TestCase):
     @patch('calico_ctl.node.check_system', autospec=True)
     @patch('calico_ctl.node._setup_ip_forwarding', autospec=True)
     @patch('calico_ctl.node._set_nf_conntrack_max', autospec=True)
-    @patch('calico_ctl.node.get_host_ips', autospec=True)
-    @patch('calico_ctl.node.warn_if_unknown_ip', autospec=True)
-    @patch('calico_ctl.node.warn_if_hostname_conflict', autospec=True)
-    @patch('calico_ctl.node.error_if_bgp_ip_conflict', autospec=True)
     @patch('calico_ctl.node.client', autospec=True)
     @patch('calico_ctl.node.docker_client', autospec=True)
     @patch('calico_ctl.node.docker', autospec=True)
@@ -188,12 +184,12 @@ class TestNode(unittest.TestCase):
     @patch('calico_ctl.node._find_or_pull_node_image', autospec=True)
     @patch('calico_ctl.node._attach_and_stream', autospec=True)
     @patch('calico_ctl.node.running_in_container', autospec=True)
-    def test_node_dockerless_start(self, m_container, m_attach_and_stream,
+    @patch('calico_ctl.node.enforce_root', autospec=True)
+    def test_node_dockerless_start(self, m_enforce_root, m_container, m_attach_and_stream,
                                    m_find_or_pull_node_image, m_call, m_docker,
                                    m_docker_client, m_client,
-                                   m_error_if_bgp_ip_conflict,
-                                   m_warn_if_hostname_conflict,
-                                   m_warn_if_unknown_ip, m_get_host_ips, m_conntrack,
+                                   
+                                   m_conntrack,
                                    m_setup_ip, m_check_system, m_os_makedirs,
                                    m_os_path_exists, m_ipv6_enabled):
         """
@@ -201,11 +197,9 @@ class TestNode(unittest.TestCase):
         without making Docker calls when runtime=none.
         """
         # Set up mock objects
+        m_enforce_root.return_value = False
         m_container.return_value = False
         m_os_path_exists.return_value = False
-        ip_1 = '1.1.1.1'
-        ip_2 = '2.2.2.2'
-        m_get_host_ips.return_value = [ip_1, ip_2]
         m_docker.utils.create_host_config.return_value = 'host_config'
         container = {'Id': 666}
         m_docker_client.create_container.return_value = container
@@ -215,7 +209,7 @@ class TestNode(unittest.TestCase):
         node_image = 'node_image'
         runtime = 'none'
         log_dir = './log_dir'
-        ip = ''
+        ip = '2.2.2.2'
         ip6 = 'aa:bb::zz'
         as_num = ''
         detach = True
@@ -234,18 +228,6 @@ class TestNode(unittest.TestCase):
                                                check_modules=True
                                                )
         m_setup_ip.assert_called_once_with()
-        m_get_host_ips.assert_called_once_with(exclude=["^docker.*", "^cbr.*",
-                                                        "virbr.*", "lxcbr.*",
-                                                        "veth.*", "cali.*",
-                                                        "tunl.*"])
-        m_warn_if_unknown_ip.assert_called_once_with(ip_2, ip6)
-        m_warn_if_hostname_conflict.assert_called_once_with(ip_2)
-        m_error_if_bgp_ip_conflict.assert_called_once_with(ip_2, ip6)
-        m_client.get_ip_pools.assert_has_calls([call(4), call(6)])
-        m_client.ensure_global_config.assert_called_once_with()
-        m_client.create_host.assert_called_once_with(
-            node.hostname, ip_2, ip6, as_num
-        )
 
         self.assertFalse(m_docker_client.remove_container.called)
         self.assertFalse(m_docker.utils.create_host_config.called)
@@ -259,27 +241,20 @@ class TestNode(unittest.TestCase):
     @patch('os.makedirs', autospec=True)
     @patch('calico_ctl.node.call', autospec=True)
     @patch('calico_ctl.node.running_in_container', autospec=True)
-    @patch('calico_ctl.node._remove_host_tunnel_addr', autospec=True)
-    @patch('calico_ctl.node._ensure_host_tunnel_addr', autospec=True)
     @patch('calico_ctl.node.check_system', autospec=True)
     @patch('calico_ctl.node._setup_ip_forwarding', autospec=True)
     @patch('calico_ctl.node._set_nf_conntrack_max', autospec=True)
-    @patch('calico_ctl.node.get_host_ips', autospec=True)
-    @patch('calico_ctl.node.warn_if_unknown_ip', autospec=True)
-    @patch('calico_ctl.node.warn_if_hostname_conflict', autospec=True)
-    @patch('calico_ctl.node.error_if_bgp_ip_conflict', autospec=True)
     @patch('calico_ctl.node.client', autospec=True)
     @patch('calico_ctl.node.docker_client', autospec=True)
     @patch('calico_ctl.node.docker', autospec=True)
     @patch('calico_ctl.node._find_or_pull_node_image', autospec=True)
     @patch('calico_ctl.node._attach_and_stream', autospec=True)
-    def test_node_start(self, m_attach_and_stream,
+    @patch('calico_ctl.node.enforce_root', autospec=True)
+    def test_node_start(self, m_root, m_attach_and_stream,
                         m_find_or_pull_node_image, m_docker,
                         m_docker_client, m_client,
-                        m_error_if_bgp_ip_conflict, m_warn_if_hostname_conflict,
-                        m_warn_if_unknown_ip, m_get_host_ips, m_conntrack, m_setup_ip,
-                        m_check_system, m_ensure_host_tunnel_addr,
-                        m_remove_host_tunnel_addr, m_container, m_call,
+                        m_conntrack, m_setup_ip,
+                        m_check_system, m_container, m_call,
                         m_os_makedirs, m_os_path_exists, m_ipv6_enabled):
         """
         Test that the node_Start function does not make Docker calls
@@ -287,10 +262,9 @@ class TestNode(unittest.TestCase):
         """
         # Set up mock objects
         m_container.return_value = False
+        m_root.return_value = False
         m_os_path_exists.return_value = False
-        ip_1 = '1.1.1.1'
         ip_2 = '2.2.2.2'
-        m_get_host_ips.return_value = [ip_1, ip_2]
         m_docker.utils.create_host_config.return_value = 'host_config'
         container = {'Id': 666}
         m_docker_client.create_container.return_value = container
@@ -304,7 +278,7 @@ class TestNode(unittest.TestCase):
         node_image = 'node_image'
         runtime = 'docker'
         log_dir = './log_dir'
-        ip = ''
+        ip = '2.2.2.2'
         ip6 = 'aa:bb::zz'
         as_num = ''
         detach = False
@@ -320,10 +294,9 @@ class TestNode(unittest.TestCase):
             "IP=%s" % ip_2,
             "IP6=%s" % ip6,
             "CALICO_NETWORKING=%s" % node.CALICO_NETWORKING_DEFAULT,
+            "AS=",
             "ETCD_AUTHORITY=%s" % ETCD_AUTHORITY_DEFAULT,  # etcd host:port
             "ETCD_SCHEME=%s" % ETCD_SCHEME_DEFAULT,
-            "FELIX_ETCDADDR=%s" % ETCD_AUTHORITY_DEFAULT,  # etcd host:port
-            "FELIX_ETCDSCHEME=%s" % ETCD_SCHEME_DEFAULT
         ]
         binds = {
             log_dir:
@@ -346,21 +319,6 @@ class TestNode(unittest.TestCase):
                                                check_docker=True,
                                                check_modules=True)
         m_setup_ip.assert_called_once_with()
-        m_get_host_ips.assert_called_once_with(exclude=["^docker.*", "^cbr.*",
-                                                        "virbr.*", "lxcbr.*",
-                                                        "veth.*", "cali.*",
-                                                        "tunl.*"])
-        m_warn_if_unknown_ip.assert_called_once_with(ip_2, ip6)
-        m_warn_if_hostname_conflict.assert_called_once_with(ip_2)
-        m_error_if_bgp_ip_conflict.assert_called_once_with(ip_2, ip6)
-        m_client.get_ip_pools.assert_has_calls([call(4), call(6)])
-        m_client.ensure_global_config.assert_called_once_with()
-        m_client.create_host.assert_called_once_with(
-            node.hostname, ip_2, ip6, as_num
-        )
-        m_ensure_host_tunnel_addr.assert_called_once_with(ipv4_pools,
-                                                          ipip_pools)
-        assert_false(m_remove_host_tunnel_addr.called)
 
         m_docker_client.remove_container.assert_called_once_with(
             'calico-node', force=True
@@ -381,21 +339,15 @@ class TestNode(unittest.TestCase):
             volumes=['/var/log/calico', "/var/run/calico"]
         )
         m_docker_client.start.assert_called_once_with(container)
-        m_attach_and_stream.assert_called_once_with(container)
+        m_attach_and_stream.assert_called_once_with(container, False)
 
     @patch('calico_ctl.node.ipv6_enabled', autospec=True, return_value=True)
     @patch('os.path.exists', autospec=True)
     @patch('os.makedirs', autospec=True)
     @patch('os.getenv', autospec=True)
-    @patch('calico_ctl.node._remove_host_tunnel_addr', autospec=True)
-    @patch('calico_ctl.node._ensure_host_tunnel_addr', autospec=True)
     @patch('calico_ctl.node.check_system', autospec=True)
     @patch('calico_ctl.node._setup_ip_forwarding', autospec=True)
     @patch('calico_ctl.node._set_nf_conntrack_max', autospec=True)
-    @patch('calico_ctl.node.get_host_ips', autospec=True)
-    @patch('calico_ctl.node.warn_if_unknown_ip', autospec=True)
-    @patch('calico_ctl.node.warn_if_hostname_conflict', autospec=True)
-    @patch('calico_ctl.node.error_if_bgp_ip_conflict', autospec=True)
     @patch('calico_ctl.node.client', autospec=True)
     @patch('calico_ctl.node.docker_client', autospec=True)
     @patch('calico_ctl.node.call', autospec=True)
@@ -403,24 +355,21 @@ class TestNode(unittest.TestCase):
     @patch('calico_ctl.node._find_or_pull_node_image', autospec=True)
     @patch('calico_ctl.node._attach_and_stream', autospec=True)
     @patch('calico_ctl.node.running_in_container', autospec=True)
-    def test_node_start_secure(self, m_container, m_attach_and_stream,
+    @patch('calico_ctl.node.enforce_root', autospec=True)
+    def test_node_start_secure(self, m_root, m_container, m_attach_and_stream,
                                m_find_or_pull_node_image, m_docker, m_call,
                                m_docker_client, m_client,
-                               m_error_if_bgp_ip_conflict,
-                               m_warn_if_hostname_conflict, m_warn_if_unknown_ip,
-                               m_get_host_ips, m_conntrack, m_setup_ip, m_check_system,
-                               m_ensure_host_tunnel_addr,
-                               m_remove_host_tunnel_addr, m_os_getenv,
-                               m_os_makedirs, m_os_path_exists, m_ipv6_enabled):
+                               m_conntrack, m_setup_ip, m_check_system,
+                               m_os_getenv, m_os_makedirs, m_os_path_exists,
+                               m_ipv6_enabled):
         """
         Test that the node_start function passes in correct values when
         secure etcd environment variables are present.
         """
         # Set up mock objects
+        m_root.return_value = False
         m_container.return_value = False
-        ip_1 = '1.1.1.1'
         ip_2 = '2.2.2.2'
-        m_get_host_ips.return_value = [ip_1, ip_2]
         container1 = {'Id': 111}
         container2 = {'Id': 222}
         m_docker_client.create_container.side_effect = iter([container1,
@@ -428,7 +377,6 @@ class TestNode(unittest.TestCase):
         m_docker.utils.create_host_config.return_value = 'host_config'
         m_os_path_exists.return_value = True
         m_check_system.return_value = [True, True, True]
-        m_client.get_ip_pools.return_value = []
 
         etcd_ca_path = "/path/to/ca.crt"
         etcd_cert_path = "/path/to/cert.crt"
@@ -448,7 +396,7 @@ class TestNode(unittest.TestCase):
         runtime = 'docker'
         log_dir = './log_dir'
         docker_plugin = "/run/docker/plugins"
-        ip = ''
+        ip = '2.2.2.2'
         ip6 = 'aa:bb::zz'
         as_num = ''
         detach = False
@@ -464,16 +412,12 @@ class TestNode(unittest.TestCase):
             "IP=%s" % ip_2,
             "IP6=%s" % ip6,
             "CALICO_NETWORKING=%s" % CALICO_NETWORKING_DEFAULT,
+            "AS=",
             "ETCD_AUTHORITY=%s" % ETCD_AUTHORITY_DEFAULT,  # etcd host:port
             "ETCD_SCHEME=%s" % "https",
             "ETCD_CA_CERT_FILE=%s" % ETCD_CA_CERT_NODE_FILE,
             "ETCD_KEY_FILE=%s" % ETCD_KEY_NODE_FILE,
             "ETCD_CERT_FILE=%s" % ETCD_CERT_NODE_FILE,
-            "FELIX_ETCDADDR=%s" % ETCD_AUTHORITY_DEFAULT,  # etcd host:port
-            "FELIX_ETCDSCHEME=https",
-            "FELIX_ETCDCAFILE=%s" % ETCD_CA_CERT_NODE_FILE,
-            "FELIX_ETCDKEYFILE=%s" % ETCD_KEY_NODE_FILE,
-            "FELIX_ETCDCERTFILE=%s" % ETCD_CERT_NODE_FILE
         ]
         environment_libnetwork = [
             "HOSTNAME=%s" % node.hostname,
@@ -508,18 +452,6 @@ class TestNode(unittest.TestCase):
                                                check_docker=True,
                                                check_modules=True)
         m_setup_ip.assert_called_once_with()
-        m_get_host_ips.assert_called_once_with(exclude=["^docker.*", "^cbr.*",
-                                                        "virbr.*", "lxcbr.*",
-                                                        "veth.*", "cali.*",
-                                                        "tunl.*"])
-        m_warn_if_unknown_ip.assert_called_once_with(ip_2, ip6)
-        m_warn_if_hostname_conflict.assert_called_once_with(ip_2)
-        m_error_if_bgp_ip_conflict.assert_called_once_with(ip_2, ip6)
-        m_client.get_ip_pools.assert_has_calls([call(4), call(6)])
-        m_client.ensure_global_config.assert_called_once_with()
-        m_client.create_host.assert_called_once_with(node.hostname, ip_2, ip6,
-                                                     as_num)
-        assert_true(m_remove_host_tunnel_addr.called)
 
         m_docker_client.remove_container.assert_has_calls([
             call('calico-node', force=True),
@@ -553,7 +485,7 @@ class TestNode(unittest.TestCase):
         ])
         m_docker_client.start.assert_has_calls([call(container1),
                                                 call(container2)])
-        m_attach_and_stream.assert_called_once_with(container1)
+        m_attach_and_stream.assert_called_once_with(container1, False)
 
 
     @patch('calico_ctl.node.ipv6_enabled', autospec=True, return_value=True)
@@ -562,24 +494,21 @@ class TestNode(unittest.TestCase):
     @patch('calico_ctl.node.check_system', autospec=True)
     @patch('calico_ctl.node._setup_ip_forwarding', autospec=True)
     @patch('calico_ctl.node._set_nf_conntrack_max', autospec=True)
-    @patch('calico_ctl.node.get_host_ips', autospec=True)
-    @patch('calico_ctl.node.warn_if_unknown_ip', autospec=True)
-    @patch('calico_ctl.node.warn_if_hostname_conflict', autospec=True)
-    @patch('calico_ctl.node.error_if_bgp_ip_conflict', autospec=True)
     @patch('calico_ctl.node.call', autospec=True)
     @patch('calico_ctl.node.client', autospec=True)
     @patch('calico_ctl.node.docker_client', autospec=True)
     @patch('calico_ctl.node.docker', autospec=True)
+    @patch('calico_ctl.node.enforce_root', autospec=True)
     def test_node_start_remove_container_error(
-            self, m_docker, m_docker_client, m_client, m_call,
-            m_error_if_bgp_ip_conflict, m_warn_if_hostname_conflict,
-            m_warn_if_unknown_ip, m_get_host_ips, m_conntrack, m_setup_ip, m_check_system,
+            self, m_root, m_docker, m_docker_client, m_client, m_call,
+            m_conntrack, m_setup_ip, m_check_system,
             m_os_makedirs, m_os_path_exists, m_ipv6_enabled):
         """
         Test that the docker client raises an APIError when it fails to
         remove a container.
         """
         # Set up mock objects
+        m_root.return_value = False
         err = APIError("Test error message", Response())
         m_docker_client.remove_container.side_effect = err
         m_check_system.return_value = [True, True, True]
@@ -604,18 +533,13 @@ class TestNode(unittest.TestCase):
     @patch('os.makedirs', autospec=True)
     @patch('calico_ctl.node.check_system', autospec=True)
     @patch('calico_ctl.node._setup_ip_forwarding', autospec=True)
-    @patch('calico_ctl.node.get_host_ips', autospec=True)
-    @patch('calico_ctl.node.warn_if_unknown_ip', autospec=True)
-    @patch('calico_ctl.node.warn_if_hostname_conflict', autospec=True)
-    @patch('calico_ctl.node.error_if_bgp_ip_conflict', autospec=True)
     @patch('calico_ctl.node.call', autospec=True)
     @patch('calico_ctl.node.client', autospec=True)
     @patch('calico_ctl.node.docker_client', autospec=True)
     @patch('calico_ctl.node.docker', autospec=True)
     def test_node_start_etcd_docker_error(
             self, m_docker, m_docker_client, m_client, m_call,
-            m_error_if_bgp_ip_conflict, m_warn_if_hostname_conflict,
-            m_warn_if_unknown_ip, m_get_host_ips, m_setup_ip, m_check_system,
+            m_setup_ip, m_check_system,
             m_os_makedirs, m_os_path_exists, m_ipv6_enabled):
         """
         Test that the node command fails if etcd or docker are not functional
@@ -645,106 +569,16 @@ class TestNode(unittest.TestCase):
                           node_image, runtime, log_dir, ip, ip6, as_num, detach,
                           libnetwork)
 
-    @patch('calico_ctl.node.ipv6_enabled', autospec=True, return_value=True)
-    @patch('sys.exit', autospec=True)
-    @patch('os.path.exists', autospec=True)
-    @patch('os.makedirs', autospec=True)
-    @patch('calico_ctl.node.check_system', autospec=True)
-    @patch('calico_ctl.node._setup_ip_forwarding', autospec=True)
-    @patch('calico_ctl.node._set_nf_conntrack_max', autospec=True)
-    @patch('calico_ctl.node.get_host_ips', autospec=True)
-    @patch('calico_ctl.node.warn_if_unknown_ip', autospec=True)
-    @patch('calico_ctl.node.warn_if_hostname_conflict', autospec=True)
-    @patch('calico_ctl.node.error_if_bgp_ip_conflict', autospec=True)
-    @patch('calico_ctl.node.call', autospec=True)
     @patch('calico_ctl.node.client', autospec=True)
     @patch('calico_ctl.node.docker_client', autospec=True)
-    @patch('calico_ctl.node.docker', autospec=True)
-    def test_node_start_no_detected_ips(
-            self, m_docker, m_docker_client, m_client, m_call,
-            m_error_if_bgp_ip_conflict, m_warn_if_hostname_conflict,
-            m_warn_if_unknown_ip, m_get_host_ips, m_conntrack, m_setup_ip, m_check_system,
-            m_os_makedirs, m_os_path_exists, m_sys_exit, m_ipv6_enabled):
-        """
-        Test that system exits when no ip is provided and host ips cannot be
-        obtained
-        """
-        # Set up mock objects
-        m_get_host_ips.return_value = []
-        m_check_system.return_value = [True, True, True]
-
-        # Set up arguments
-        node_image = 'node_image'
-        runtime = 'docker'
-        log_dir = './log_dir'
-        ip = ''
-        ip6 = 'aa:bb::zz'
-        as_num = ''
-        detach = False
-        libnetwork = False
-
-        # Call method under test
-        node.node_start(node_image, runtime, log_dir, ip, ip6, as_num, detach,
-                        libnetwork)
-
-        # Assert
-        m_sys_exit.assert_called_once_with(1)
-
-    @patch('calico_ctl.node.ipv6_enabled', autospec=True, return_value=True)
-    @patch('os.path.exists', autospec=True)
-    @patch('os.makedirs', autospec=True)
-    @patch('calico_ctl.node.check_system', autospec=True)
-    @patch('calico_ctl.node._setup_ip_forwarding', autospec=True)
-    @patch('calico_ctl.node._set_nf_conntrack_max', autospec=True)
-    @patch('calico_ctl.node.get_host_ips', autospec=True)
-    @patch('calico_ctl.node.warn_if_unknown_ip', autospec=True)
-    @patch('calico_ctl.node.warn_if_hostname_conflict', autospec=True)
-    @patch('calico_ctl.node.error_if_bgp_ip_conflict', autospec=True)
-    @patch('calico_ctl.node.call', autospec=True)
-    @patch('calico_ctl.node.client', autospec=True)
-    @patch('calico_ctl.node.docker_client', autospec=True)
-    @patch('calico_ctl.node.docker', autospec=True)
-    def test_node_start_create_default_ip_pools(
-            self, m_docker, m_docker_client, m_client, m_call,
-            m_error_if_bgp_ip_conflict, m_warn_if_hostname_conflict,
-            m_warn_if_unknown_ip, m_get_host_ips, m_conntrack,  m_setup_ip, m_check_system,
-            m_os_makedirs, m_os_path_exists, m_ipv6_enabled):
-        """
-        Test that the client creates default ipv4 and ipv6 pools when the
-        client returns an empty ip_pool on etcd setup
-        """
-        # Set up mock objects
-        m_client.get_ip_pools.return_value = []
-        m_check_system.return_value = [True, True, True]
-
-        # Set up arguments
-        node_image = 'node_image'
-        runtime = 'docker'
-        log_dir = './log_dir'
-        ip = ''
-        ip6 = 'aa:bb::zz'
-        as_num = ''
-        detach = False
-        libnetwork = False
-
-        # Call method under test
-        node.node_start(node_image, runtime, log_dir, ip, ip6, as_num, detach,
-                        libnetwork)
-
-        # Assert
-        m_client.add_ip_pool.assert_has_calls([
-            call(4, node.DEFAULT_IPV4_POOL),
-            call(6, node.DEFAULT_IPV6_POOL)
-        ])
-
-    @patch('calico_ctl.node.client', autospec=True)
-    @patch('calico_ctl.node.docker_client', autospec=True)
-    def test_node_stop(self, m_docker_client, m_client):
+    @patch('calico_ctl.node.enforce_root', autospec=True)
+    def test_node_stop(self, m_root, m_docker_client, m_client):
         """
         Test the client stops the node when node_stop called when there are
         endpoints and the force flag is set.
         """
         # Call method under test
+        m_root.return_value = False
         m_client.get_endpoints.return_value = [Mock()]
         node.node_stop(True)
 
@@ -755,12 +589,14 @@ class TestNode(unittest.TestCase):
 
     @patch('calico_ctl.node.client', autospec=True)
     @patch('calico_ctl.node.docker_client', autospec=True)
-    def test_node_stop_endpoints(self, m_docker_client, m_client):
+    @patch('calico_ctl.node.enforce_root', autospec=True)
+    def test_node_stop_endpoints(self, m_root, m_docker_client, m_client):
         """
         Test the client does not stops the node when node_stop is called and
         there are endpoints and the force flag is not set.
         """
         # Call method under test
+        m_root.return_value = False
         m_client.get_endpoints.return_value = [Mock()]
         self.assertRaises(SystemExit, node.node_stop, False)
 
@@ -770,12 +606,14 @@ class TestNode(unittest.TestCase):
 
     @patch('calico_ctl.node.client', autospec=True)
     @patch('calico_ctl.node.docker_client', autospec=True)
-    def test_node_stop_error(self, m_docker_client, m_client):
+    @patch('calico_ctl.node.enforce_root', autospec=True)
+    def test_node_stop_error(self, m_root, m_docker_client, m_client):
         """
         Test node_stop raises an exception when the docker client cannot not
         stop the node
         """
         # Set up mock objects
+        m_root.return_value = False
         m_client.get_endpoints.return_value = [Mock()]
         err = APIError("Test error message", Response())
 
@@ -785,12 +623,10 @@ class TestNode(unittest.TestCase):
             # Call method under test expecting an exception
             self.assertRaises(APIError, node.node_stop, True)
 
-    @patch('calico_ctl.node._remove_host_tunnel_addr', autospec=True)
     @patch('calico_ctl.node.remove_veth', autospec=True)
     @patch('calico_ctl.node._container_running', autospec=True, return_value=False)
     @patch('calico_ctl.node.client', autospec=True)
-    def test_node_remove(self, m_client, m_cont_running, m_veth,
-                         m_remove_tunnel_addr):
+    def test_node_remove(self, m_client, m_cont_running, m_veth):
         """
         Test the client removes the host when node_remove called, and that
         endpoints are removed when remove_endpoints flag is set.
@@ -808,7 +644,6 @@ class TestNode(unittest.TestCase):
         m_client.remove_host.assert_called_once_with(node.hostname)
         m_veth.assert_has_calls([call("vethname1"), call("vethname2")])
         m_cont_running.assert_has_calls([call("calico-node"), call("calico-libnetwork")])
-        assert_equal(m_remove_tunnel_addr.mock_calls, [call()])
 
     @patch('calico_ctl.node.remove_veth', autospec=True)
     @patch('calico_ctl.node._container_running', autospec=True, return_value=True)
@@ -873,141 +708,3 @@ class TestNode(unittest.TestCase):
         for test in (True, False):
             m_docker_client.inspect_container.return_value = {"State": {"Running": test}}
             self.assertEquals(node._container_running("container1"), test)
-
-
-    @patch("calico_ctl.node.client", autospec=True)
-    @patch("sys.exit", autospec=True)
-    def test_error_if_bgp_ipv4_conflict_no_conflict(self, m_exit, m_client):
-        """
-        Test check that node IP is not already in use by another node, no error.
-        """
-        m_client.get_hostnames_from_ips = Mock()
-        m_client.get_hostnames_from_ips.return_value = {}
-        node.error_if_bgp_ip_conflict("10.0.0.1", "abcd::beef")
-        self.assertFalse(m_exit.called)
-
-    @patch("calico_ctl.node.client", autospec=True)
-    @patch("calico_ctl.utils.get_hostname", autospec=True)
-    @patch("sys.exit", autospec=True)
-    def test_error_if_ip_conflict_ipv6_key_error(self, m_exit, m_hostname,
-                                                 m_client):
-        """
-        Test that function accepts IP being owned by same host.
-        """
-        calico_ctl.node.hostname = "host"
-        m_client.get_hostnames_from_ips = Mock()
-        m_client.get_hostnames_from_ips.return_value = {"10.0.0.1":"host"}
-        node.error_if_bgp_ip_conflict("10.0.0.1", "abcd::beef")
-        self.assertFalse(m_exit.called)
-
-    @patch("calico_ctl.node.client", autospec=True)
-    @patch("calico_ctl.utils.get_hostname", autospec=True)
-    def test_error_when_bgp_ipv4_conflict(self, m_hostname, m_client):
-        """
-        Test that function exits when another node already uses ipv4 addr.
-        """
-        calico_ctl.node.hostname = "not_host"
-        m_client.get_hostnames_from_ips = Mock()
-        m_client.get_hostnames_from_ips.return_value = {"10.0.0.1":"host"}
-        self.assertRaises(SystemExit, node.error_if_bgp_ip_conflict,
-                          "10.0.0.1", None)
-
-    @patch("calico_ctl.node.client", autospec=True)
-    @patch("calico_ctl.utils.get_hostname", autospec=True)
-    def test_error_when_bgp_ipv6_conflict(self, m_hostname, m_client):
-        """
-        Test that function exits when another node already uses ipv6 addr.
-        """
-        calico_ctl.node.hostname = "not_host"
-        m_client.get_hostnames_from_ips = Mock()
-        m_client.get_hostnames_from_ips.return_value = {"abcd::beef":"host"}
-        self.assertRaises(SystemExit, node.error_if_bgp_ip_conflict,
-                          None, "abcd::beef")
-
-    @patch("calico_ctl.node._get_host_tunnel_ip", autospec=True)
-    @patch("calico_ctl.node._assign_host_tunnel_addr", autospec=True)
-    @patch("calico_ctl.node.client", autospec=True)
-    @patch("calico_ctl.utils.get_hostname", autospec=True)
-    def test_ensure_host_tunnel_addr_no_ip(self, m_hostname, m_client,
-                                           m_assign_host_tunnel_addr,
-                                           m_get_tunnel_host_ip):
-        m_get_tunnel_host_ip.return_value = None
-        ipv4_pools = [IPPool("10.0.0.0/16"),
-                      IPPool("10.1.0.0/16", ipip=True)]
-        ipip_pools = [IPPool("10.1.0.0/16", ipip=True)]
-        calico_ctl.node._ensure_host_tunnel_addr(ipv4_pools, ipip_pools)
-        assert_equal(m_assign_host_tunnel_addr.mock_calls, [call(ipip_pools)])
-
-    @patch("calico_ctl.node._get_host_tunnel_ip", autospec=True)
-    @patch("calico_ctl.node._assign_host_tunnel_addr", autospec=True)
-    @patch("calico_ctl.node.client", autospec=True)
-    @patch("calico_ctl.utils.get_hostname", autospec=True)
-    def test_ensure_host_tunnel_addr_non_ipip(self, m_hostname, m_client,
-                                              m_assign_host_tunnel_addr,
-                                              m_get_tunnel_host_ip):
-        m_get_tunnel_host_ip.return_value = IPAddress("10.0.0.1")
-        ipv4_pools = [IPPool("10.0.0.0/16"),
-                      IPPool("10.1.0.0/16", ipip=True)]
-        ipip_pools = [IPPool("10.1.0.0/16", ipip=True)]
-        calico_ctl.node._ensure_host_tunnel_addr(ipv4_pools, ipip_pools)
-        assert_equal(m_client.release_ips.mock_calls,
-                     [call({IPAddress("10.0.0.1")})])
-        assert_equal(m_assign_host_tunnel_addr.mock_calls, [call(ipip_pools)])
-
-    @patch("calico_ctl.node._get_host_tunnel_ip", autospec=True)
-    @patch("calico_ctl.node._assign_host_tunnel_addr", autospec=True)
-    @patch("calico_ctl.node.client", autospec=True)
-    @patch("calico_ctl.utils.get_hostname", autospec=True)
-    def test_ensure_host_tunnel_addr_bad_ip(self, m_hostname, m_client,
-                                            m_assign_host_tunnel_addr,
-                                            m_get_tunnel_host_ip):
-        m_get_tunnel_host_ip.return_value = IPAddress("11.0.0.1")
-        ipv4_pools = [IPPool("10.0.0.0/16"),
-                      IPPool("10.1.0.0/16", ipip=True)]
-        ipip_pools = [IPPool("10.1.0.0/16", ipip=True)]
-        calico_ctl.node._ensure_host_tunnel_addr(ipv4_pools, ipip_pools)
-        assert_equal(m_assign_host_tunnel_addr.mock_calls, [call(ipip_pools)])
-
-    @patch("calico_ctl.node.client", autospec=True)
-    @patch("calico_ctl.node.hostname", autospec=True)
-    def test_assign_host_tunnel_addr(self, m_hostname, m_client):
-        # First pool full, IP allocated from second pool.
-        m_client.auto_assign_ips.side_effect = iter([
-            ([], []),
-            ([IPAddress("10.0.0.1")], [])
-        ])
-        ipip_pools = [IPPool("10.1.0.0/16", ipip=True),
-                      IPPool("10.0.0.0/16", ipip=True)]
-        calico_ctl.node._assign_host_tunnel_addr(ipip_pools)
-        assert_equal(
-            m_client.set_per_host_config.mock_calls,
-            [call(m_hostname, "IpInIpTunnelAddr", "10.0.0.1")]
-        )
-
-    @patch("sys.exit", autospec=True)
-    @patch("calico_ctl.node.client", autospec=True)
-    @patch("calico_ctl.node.hostname", autospec=True)
-    def test_assign_host_tunnel_addr_none_available(self, m_hostname,
-                                                    m_client, m_exit):
-        # First pool full, IP allocated from second pool.
-        m_client.auto_assign_ips.side_effect = iter([
-            ([], []),
-            ([], [])
-        ])
-        ipip_pools = [IPPool("10.1.0.0/16", ipip=True),
-                      IPPool("10.0.0.0/16", ipip=True)]
-        m_exit.side_effect = Exception
-        assert_raises(Exception, calico_ctl.node._assign_host_tunnel_addr,
-                      ipip_pools)
-        assert_equal(m_exit.mock_calls, [call(1)])
-
-    @patch("calico_ctl.node._get_host_tunnel_ip", autospec=True)
-    @patch("calico_ctl.node.client", autospec=True)
-    @patch("calico_ctl.node.hostname", autospec=True)
-    def test_remove_host_tunnel_addr(self, m_hostname, m_client, m_get_ip):
-        ip_address = IPAddress("10.0.0.1")
-        m_get_ip.return_value = ip_address
-        calico_ctl.node._remove_host_tunnel_addr()
-        assert_equal(m_client.release_ips.mock_calls, [call({ip_address})])
-        assert_equal(m_client.remove_per_host_config.mock_calls,
-                     [call(m_hostname, "IpInIpTunnelAddr")])
